% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{abortifnot}
\alias{abortifnot}
\title{Abort if not}
\usage{
abortifnot(cnd, message = NULL)
}
\arguments{
\item{cnd}{Condition to be evaluated.}

\item{message}{Message to display if condition is false.}
}
\description{
Abort if not
}
\author{
João Macalós
}
\keyword{internal}
