% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_baseline.R
\name{.abort_if_dup}
\alias{.abort_if_dup}
\title{Abort if duplicated variables}
\usage{
.abort_if_dup(dups)
}
\arguments{
\item{dups}{name(s) of offending variables}
}
\description{
Abort if duplicated variables
}
\author{
João Macalós
}
\keyword{internal}
