% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_matrix.R
\name{.abort_typo_code}
\alias{.abort_typo_code}
\title{Abort if typo on the codes of columns}
\usage{
.abort_typo_code(nms)
}
\arguments{
\item{nms}{Incorrect codes detected}
}
\description{
Abort if typo on the codes of columns
}
\author{
João Macalós
}
\keyword{internal}
