% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_validate.R
\name{.abort_water_leakc}
\alias{.abort_water_leakc}
\title{Abort if column validation is not fulfilled}
\usage{
.abort_water_leakc(c2names, which)
}
\arguments{
\item{c2names}{Names of offending columns}

\item{which}{Balance-sheet or transactions-flow matrix?}
}
\description{
Abort if column validation is not fulfilled
}
\author{
João Macalós
}
\keyword{internal}
