% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_validate.R
\name{.abort_water_leakr}
\alias{.abort_water_leakr}
\title{Abort if row validation is not fulfilled}
\usage{
.abort_water_leakr(r2names, which)
}
\arguments{
\item{r2names}{Names of offending rows}

\item{which}{Balance-sheet or transactions-flow matrix?}
}
\description{
Abort if row validation is not fulfilled
}
\author{
João Macalós
}
\keyword{internal}
