% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_find_order.R
\name{.add_time2}
\alias{.add_time2}
\title{Find dependencies and order the equations}
\usage{
.add_time2(x)
}
\arguments{
\item{x}{A vector to modify}
}
\description{
Find dependencies and order the equations
}
\author{
João Macalós
}
\keyword{internal}
