% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_find_order.R
\name{.add_time_stamps}
\alias{.add_time_stamps}
\title{Find dependencies and order the equations}
\usage{
.add_time_stamps(eq_as_tb)
}
\arguments{
\item{eq_as_tb}{A tibble generated with \code{.eq_as_tb()}.}
}
\description{
Find dependencies and order the equations
}
\author{
João Macalós
}
