% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_validate.R
\name{.all_equal}
\alias{.all_equal}
\title{Check if all values in x are equal}
\usage{
.all_equal(x, tol)
}
\arguments{
\item{x}{A numeric vector}

\item{tol}{Tolerance to declare equality}
}
\description{
Check if all values in x are equal
}
\author{
João Macalós
}
\keyword{internal}
