% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_matrix.R
\name{.args_to_row}
\alias{.args_to_row}
\title{Take arguments and make them a row of a tibble}
\usage{
.args_to_row(arg, tb)
}
\arguments{
\item{arg}{The arguments to transform}

\item{tb}{The Tibble that will receive the rows}
}
\description{
Take arguments and make them a row of a tibble
}
\author{
João Macalós
}
\keyword{internal}
