% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_portfolio.R
\name{.check_symmetry}
\alias{.check_symmetry}
\title{Check that symmetry condition is valid and fulfill missing entries}
\usage{
.check_symmetry(m)
}
\arguments{
\item{m}{A square matrix}
}
\description{
Check that symmetry condition is valid and fulfill missing entries
}
\author{
JOão Macalós
}
\keyword{internal}
