% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_find_order.R
\name{.eq_as_tb}
\alias{.eq_as_tb}
\title{Split the formulae into a \code{tibble} with the left-hand side and
right-hand side as columns}
\usage{
.eq_as_tb(equations)
}
\arguments{
\item{equations}{list of equations}
}
\description{
Split the formulae into a \code{tibble} with the left-hand side and
right-hand side as columns
}
\author{
João Macalós
}
\keyword{internal}
