% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_scenario.R
\name{.extend_baseline_matrix}
\alias{.extend_baseline_matrix}
\title{Extend a baseline matrix}
\usage{
.extend_baseline_matrix(baseline, periods)
}
\arguments{
\item{baseline}{A baseline model}

\item{periods}{The total number of periods to run the model}
}
\description{
This function is called if a scenario is to be created that just
continues with the baseline specification. It is useful to create
a benchmark model to compare new scenarios.
}
