% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_portfolio.R
\name{.fill_rows_and_cols}
\alias{.fill_rows_and_cols}
\title{Fill all possible rows and columns and validate}
\usage{
.fill_rows_and_cols(m)
}
\arguments{
\item{m}{A square matrix}
}
\description{
Fill all possible rows and columns and validate
}
\author{
João Macalós
}
\keyword{internal}
