% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_find_order.R
\name{.find_blocks}
\alias{.find_blocks}
\title{Find blocks of independent equations}
\usage{
.find_blocks(adj)
}
\arguments{
\item{adj}{Adjacency matrix}
}
\description{
Find blocks of independent equations
}
