% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_matrix_display.R
\name{.find_names}
\alias{.find_names}
\title{Find names for display matrix}
\usage{
.find_names(matrix)
}
\arguments{
\item{matrix}{A balance-sheet or transactions-flow matrix}
}
\description{
Clean cells to display in latex format
}
\author{
João Macalós
}
\keyword{internal}
