% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_validate.R
\name{.get_matrix}
\alias{.get_matrix}
\title{Get numeric matrix for evaluation from balance-sheet or
transactions-flow matrices}
\usage{
.get_matrix(mtrx, bl1, bl2)
}
\arguments{
\item{mtrx}{Balance-sheet or transactions-flow matrix}

\item{bl1}{calls from baseline model}

\item{bl2}{external from baseline model}
}
\description{
Get numeric matrix for evaluation from balance-sheet or
transactions-flow matrices
}
\author{
João Macalós
}
\keyword{internal}
