% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_validate.R
\name{.is_equal}
\alias{.is_equal}
\title{Check if two values are equal}
\usage{
.is_equal(x, y, tol)
}
\arguments{
\item{x, y}{numeric values}

\item{tol}{Tolerance to declare equality}
}
\description{
Check if two values are equal
}
\author{
João Macalós
}
\keyword{internal}
