% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_is_hidden_true.R
\name{.is_under_tol}
\alias{.is_under_tol}
\title{Check if difference is below a given tolerance level}
\usage{
.is_under_tol(x, y, tol = 0.1)
}
\arguments{
\item{y, x}{Numeric values}

\item{tol}{Numeric value determining the tolerance for the test}
}
\description{
Check if difference is below a given tolerance level
}
