% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-solvers.R
\name{.make_matrix}
\alias{.make_matrix}
\title{Make the underlying matrix that will be modified in place with the Gauss Seidel
algorithm}
\usage{
.make_matrix(equations, external, periods, initial = NULL)
}
\arguments{
\item{equations}{Prepared equations.}

\item{external}{Exogenous and parameters as tibble.}

\item{periods}{Total number of rows.}

\item{initial}{Initial values, if supplied.}
}
\description{
Make the underlying matrix that will be modified in place with the Gauss Seidel
algorithm
}
\author{
João Macalós
}
