% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-solvers.R
\name{.prep_broyden}
\alias{.prep_broyden}
\title{Prep equations for Broyden and Newton solvers}
\usage{
.prep_broyden(.block)
}
\arguments{
\item{.block}{Blocks of equations}
}
\description{
Prep equations for Broyden and Newton solvers
}
\author{
João Macalós
}
\keyword{internal}
