% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-solvers.R
\name{.prep_equations}
\alias{.prep_equations}
\title{Re-wrote the equations with the correct matrix syntax that will be used to evaluate
the expressions inside the Gauss Seidel algorithm}
\usage{
.prep_equations(ordered_eqs, external)
}
\arguments{
\item{ordered_eqs}{ordered equations after passing through \code{.sfcr_find_order()} function.}

\item{external}{Tibble of exogenous values and parameters, already separated with
\code{.eq_as_tb()} function.}
}
\description{
Re-wrote the equations with the correct matrix syntax that will be used to evaluate
the expressions inside the Gauss Seidel algorithm
}
\author{
João Macalós
}
