% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_find_order.R
\name{.pvar}
\alias{.pvar}
\title{Pattern replacement var}
\usage{
.pvar(x)
}
\arguments{
\item{x}{vector of variables}
}
\description{
Pattern replacement var
}
\author{
João Macalós
}
\keyword{internal}
