% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_find_order.R
\name{.pvarlag}
\alias{.pvarlag}
\title{Pattern replacement lag}
\usage{
.pvarlag(x)
}
\arguments{
\item{x}{vector of variables}
}
\description{
Pattern replacement lag
}
\author{
João Macalós
}
\keyword{internal}
