% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_plots.R
\name{.return_loops}
\alias{.return_loops}
\title{Find cyclical nodes}
\usage{
.return_loops(m)
}
\arguments{
\item{m}{adjacency matrix}
}
\description{
Find cyclical nodes
}
\note{
See Networks: an introduction from M.E.J. Newman, 2010, p. 136-139
for a reference on this algorithm.
}
\keyword{internal}
