% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_portfolio.R
\name{.scan_cols}
\alias{.scan_cols}
\title{Scan columns to fill whenever there's only one value missing.}
\usage{
.scan_cols(m)
}
\arguments{
\item{m}{A square matrix}
}
\description{
Scan columns to fill whenever there's only one value missing.
}
\author{
João Macalós
}
\keyword{internal}
