% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_portfolio.R
\name{.scan_rows}
\alias{.scan_rows}
\title{Scan rows to fill whenever there's only one value missing.}
\usage{
.scan_rows(m)
}
\arguments{
\item{m}{A square matrix}
}
\description{
Scan rows to fill whenever there's only one value missing.
}
\author{
João Macalós
}
\keyword{internal}
