% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_broyden.R
\name{.sfcr_broyden}
\alias{.sfcr_broyden}
\title{Broyden solver wrapper}
\usage{
.sfcr_broyden(m, equations, periods, max_ite, tol)
}
\arguments{
\item{m}{The initialized matrix obtained with \code{.make_matrix()}.}

\item{equations}{Prepared equations with \code{.prep_equations()}.}

\item{periods}{Total number of rows (periods) in the model.}

\item{max_ite}{Maximum number of iterations allowed per block per period.}

\item{tol}{Tolerance accepted to determine convergence.}
}
\description{
Broyden solver wrapper
}
\details{
This function implements the Broyden method to solve the cyclical
blocks of equations.
}
\author{
João Macalós
}
\keyword{internal}
