% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_baseline.R
\name{.sfcr_eqs_check}
\alias{.sfcr_eqs_check}
\title{Check for missing endogenous variables}
\usage{
.sfcr_eqs_check(m, equations)
}
\arguments{
\item{m}{The initialized matrix obtained with \code{.make_matrix()}.}

\item{equations}{Prepared equations with \code{.prep_equations()}.}
}
\description{
Check for missing endogenous variables
}
\author{
João Macalós
}
