% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_find_order.R
\name{.sfcr_find_adjacency}
\alias{.sfcr_find_adjacency}
\title{Find adjacency matrix for a system of equations}
\usage{
.sfcr_find_adjacency(equations)
}
\arguments{
\item{equations}{A system of equations already time stamped}
}
\description{
Find adjacency matrix for a system of equations
}
\author{
João Macalós
}
\keyword{internal}
