% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_find_order.R
\name{.sfcr_find_order}
\alias{.sfcr_find_order}
\title{Place the equations in the correct order for estimation}
\usage{
.sfcr_find_order(equations)
}
\arguments{
\item{equations}{Equations supplied by the user.}
}
\description{
Place the equations in the correct order for estimation
}
\details{
Create an adjacency matrix and apply
\code{.find_blocks()} function to identify the blocks
of independent equations.
}
\author{
João Macalós
}
