% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_is_hidden_true.R
\name{.sfcr_is_hidden_true}
\alias{.sfcr_is_hidden_true}
\title{Check if hidden equation condition is validated}
\usage{
.sfcr_is_hidden_true(m, hidden, .hidden_tol)
}
\arguments{
\item{m}{A matrix simulated with \code{sfcr_gauss_seidel()}.}

\item{hidden}{A named list containing the hidden equation.}

\item{.hidden_tol}{The tolerance to accept equality in the hidden equation}
}
\description{
Check if hidden equation condition is validated
}
