% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_matrix_display.R
\name{.to_latex_style}
\alias{.to_latex_style}
\title{Transform entries into latex style}
\usage{
.to_latex_style(matrix, nms)
}
\arguments{
\item{matrix}{a balance-sheet or transactions-flow matrix}

\item{nms}{Cleaned names with \code{.find_names()}}
}
\description{
Transform entries into latex style
}
\author{
João Macalós
}
\keyword{internal}
