% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_validate.R
\name{.validate_matrix}
\alias{.validate_matrix}
\title{Validate a balance-sheet or transactions-flow matrix on the simulated data}
\usage{
.validate_matrix(mtrx, m, which = "tfm", tol, rtol = FALSE)
}
\arguments{
\item{mtrx}{A balance-sheet or transactions-flow matrix}

\item{m}{A baseline model in matrix format -- get from attributes or with
\code{sfcr_get_matrix()} function.}

\item{which}{A balance-sheet or a transactions-flow matrix?}

\item{tol}{Tolerance for convergence}

\item{rtol}{Relative tolerance?}
}
\description{
Validate a balance-sheet or transactions-flow matrix on the simulated data
}
\author{
João Macalós
}
\keyword{internal}
