% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_portfolio.R
\name{.validate_scan}
\alias{.validate_scan}
\title{Check that the portfolio matrix respect the horizontal and vertical
adding-up constraints}
\usage{
.validate_scan(m, which)
}
\arguments{
\item{m}{A square matrix}

\item{which}{Rows or columns?}
}
\description{
Check that the portfolio matrix respect the horizontal and vertical
adding-up constraints
}
\author{
João Macalós
}
\keyword{internal}
