% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_multis.R
\name{new_sfcr_mlt}
\alias{new_sfcr_mlt}
\title{sfcr_mlt constructor}
\usage{
new_sfcr_mlt(multis, fixed)
}
\arguments{
\item{multis}{A list with multiple \code{sfcr_tbl}}

\item{fixed}{The fixed argument}
}
\description{
sfcr_mlt constructor
}
\author{
João Macalós
}
\keyword{internal}
