% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_expand.R
\name{new_sfcr_mlt_set}
\alias{new_sfcr_mlt_set}
\title{sfcr_mlt_set constructor}
\usage{
new_sfcr_mlt_set(x, original)
}
\arguments{
\item{x}{A expanded list of sfcr_set's}

\item{original}{the original sfcr_set}
}
\description{
sfcr_mlt_set constructor
}
\author{
João Macalós
}
\keyword{internal}
