% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_expand.R
\name{new_sfcr_mlt_shock}
\alias{new_sfcr_mlt_shock}
\title{sfcr_mlt_shock constructor}
\usage{
new_sfcr_mlt_shock(x, original)
}
\arguments{
\item{x}{A expanded list of sfcr_shock's}

\item{original}{the original sfcr_shock}
}
\description{
sfcr_mlt_shock constructor
}
\author{
João Macalós
}
\keyword{internal}
