% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_set.R
\name{new_sfcr_set}
\alias{new_sfcr_set}
\title{sfcr_set constructor}
\usage{
new_sfcr_set(list)
}
\arguments{
\item{list}{A list}
}
\description{
sfcr_set constructor
}
\author{
João Macalós
}
\keyword{internal}
