% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_baseline.R
\name{new_sfcr_tbl}
\alias{new_sfcr_tbl}
\title{new_sfcr_tbl constructor}
\usage{
new_sfcr_tbl(tbl, matrix, calls, external)
}
\arguments{
\item{tbl}{A tibble}

\item{matrix}{a Matrix}

\item{calls}{Calls tibble}

\item{external}{Vector with external names}
}
\description{
new_sfcr_tbl constructor
}
\author{
João Macalós
}
\keyword{internal}
