% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_plots.R
\name{sfcr_dag_blocks}
\alias{sfcr_dag_blocks}
\title{Create a \code{tbl_graph} object blocks and cycles information}
\usage{
sfcr_dag_blocks(equations)
}
\arguments{
\item{equations}{A list containing all the equations of the model to be simulated. The equations
must be written in R syntax, with the left-hand side separated from the right-hand side
by a tilde \code{~}.}
}
\value{
A \code{tbl_graph}
}
\description{
Create a \code{tbl_graph} object blocks and cycles information
}
\details{
This function creates a \code{tbl_graph} with information about
the blocks and cycles attached to it. This object can then be used to
plot the DAG of the model.
}
\author{
João Macalós, \email{joaomacalos@gmail.com}
}
