% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_plots.R
\name{sfcr_dag_blocks_plot}
\alias{sfcr_dag_blocks_plot}
\title{Plot the DAG with blocks and cycles information}
\usage{
sfcr_dag_blocks_plot(equations, title = NULL, size = 10)
}
\arguments{
\item{equations}{A \code{sfcr_set} containing all the equations of the model to be simulated. The equations
must be written with the R formula syntax, with the left-hand side separated from the right-hand side
by a twiddle \code{~}.}

\item{title}{Title of the plot.}

\item{size}{Size of the points.}
}
\value{
A \code{tbl_graph}
}
\description{
Plot the DAG with blocks and cycles information
}
\details{
This function creates a \code{tbl_graph} with information about
the cycles attached to it. This object can then be used to
plot the DAG of the model.
}
\author{
João Macalós, \email{joaomacalos@gmail.com}
}
