% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_plots.R
\name{sfcr_dag_cycles}
\alias{sfcr_dag_cycles}
\title{Create a \code{tbl_graph} object with cycles information}
\usage{
sfcr_dag_cycles(equations)
}
\arguments{
\item{equations}{A \code{sfcr_set} containing all the equations of the model to be simulated. The equations
must be written with the R formula syntax, with the left-hand side separated from the right-hand side
by a twiddle \code{~}.}
}
\value{
A \code{tbl_graph}
}
\description{
Create a \code{tbl_graph} object with cycles information
}
\details{
This function creates a \code{tbl_graph} with information about
the cycles attached to it. This object can then be used to
plot the DAG of the model.
}
\author{
João Macalós, \email{joaomacalos@gmail.com}
}
