% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_baseline.R
\name{sfcr_get_blocks}
\alias{sfcr_get_blocks}
\title{Get block structure of a \code{sfcr_tbl} object}
\usage{
sfcr_get_blocks(sfcr_tbl)
}
\arguments{
\item{sfcr_tbl}{A \code{sfcr_tbl} object.}
}
\description{
Get block structure of a \code{sfcr_tbl} object
}
\author{
João Macalós
}
