% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_baseline.R
\name{sfcr_get_matrix}
\alias{sfcr_get_matrix}
\title{Get Matrix form of \code{sfcr_tbl} object}
\usage{
sfcr_get_matrix(sfcr_tbl)
}
\arguments{
\item{sfcr_tbl}{A \code{sfcr_tbl} object.}
}
\description{
Get Matrix form of \code{sfcr_tbl} object
}
