% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_sankey.R
\name{sfcr_sankey}
\alias{sfcr_sankey}
\title{Plot Sankey's diagram representation of transactions-flow matrix}
\usage{
sfcr_sankey(tfm, baseline, when = "start")
}
\arguments{
\item{tfm}{A transactions-flow matrix}

\item{baseline}{A baseline model}

\item{when}{When the Sankey's diagram should be evaluated?
\itemize{
\item "start": Fifth and fourth periods.
\item "end": last two periods of the simulation (stationary state).
}}
}
\description{
Plot Sankey's diagram representation of transactions-flow matrix
}
\author{
João Macalós
}
