% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_set.R
\name{sfcr_set.formula}
\alias{sfcr_set.formula}
\title{S3 method for sfcr_set}
\usage{
\method{sfcr_set}{formula}(..., exclude = NULL)
}
\arguments{
\item{...}{The formulas used to define the equations and external
values of the system}

\item{exclude}{One or more indices of equations to be excluded. The
correct indices can be found with \code{sfcr_set_index()}.}
}
\description{
S3 method for sfcr_set
}
\author{
João Macalós
}
