% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_set.R
\name{sfcr_set.sfcr_set}
\alias{sfcr_set.sfcr_set}
\title{S3 method for sfcr_set}
\usage{
\method{sfcr_set}{sfcr_set}(..., exclude = NULL)
}
\arguments{
\item{...}{A sfcr_set and new formulas to be added}

\item{exclude}{One or more indices of equations to be excluded. The
correct indices can be found with \code{sfcr_set_index()}.}
}
\description{
S3 method for sfcr_set
}
\author{
João Macalós
}
