% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_set.R
\name{sfcr_set_index}
\alias{sfcr_set_index}
\title{Get names of endogenous vars and their index}
\usage{
sfcr_set_index(eqs)
}
\arguments{
\item{eqs}{A list of equations created with \code{sfcr_set()}}
}
\description{
The \code{sfcr_set_index()} function takes a list of equations as its input and returns
a tibble containing the name of the variable on the left-hand side of the equations
and their position in the equations list.
}
\details{
This function aims to facilitate locating a specific equation in the list in order to
modify the list of equations.
}
\author{
João Macalós
}
