% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_matrix.R
\name{sfcr_validate}
\alias{sfcr_validate}
\title{Validate a transactions-flow or balance-sheet matrix}
\usage{
sfcr_validate(matrix, baseline, which = "tfm")
}
\arguments{
\item{matrix}{A transactions-flow or balance sheet matrix}

\item{baseline}{A baseline model.}

\item{which}{Is it a transactions-flow or a balance-sheet matrix?}
}
\description{
This function validates a transactions-flow or balance-sheet
matrix with the simulated data obtained with \code{sfcr_baseline()}
function
}
