% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_broyden.R
\name{.broyden_solver}
\alias{.broyden_solver}
\title{Broyden solver}
\usage{
.broyden_solver(.x0, .fn, max_ite, tol)
}
\arguments{
\item{.x0}{Vector with initial guess for x.}

\item{.fn}{A function containing the system of equations.}

\item{max_ite}{Maximum number of iterations allowed}

\item{tol}{A numeric value indicating the accepted tolerance to declare convergence.}
}
\description{
Broyden solver
}
\author{
João Macalós
}
