% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_find_order.R
\name{.find_blocks}
\alias{.find_blocks}
\title{Algorithm to identify the blocks of independent equations inside the model.}
\usage{
.find_blocks(m)
}
\arguments{
\item{m}{A adjacency matrix of the endogenous variables.}
}
\description{
Algorithm to identify the blocks of independent equations inside the model.
}
\details{
This algorithm finds the block of independent equations sequentially.
It first looks for all variables that depends only on exogenous variables or
on lagged values. It saves these variables as the first block and eliminate
them from the adjacency matrix.
It repeats this process until all blocks are identified or until it finds
a cycle.
If a cycle is found, it jumps to the bottom and finds all variables that
does not have any children. The algorithm assign these variables to the
end of the block list and eliminates them sequentially until it reaches
the cycle block.

All the cyclical variables are treated as a single cycle. Hence, this algorithm
is unable to identify independent cycles. However, this is not a common
structure of SFC models.
}
\author{
João Macalós
}
