% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_gauss_seidel.R
\name{.sfcr_gauss_seidel}
\alias{.sfcr_gauss_seidel}
\title{Gauss Seidel algorithm}
\usage{
.sfcr_gauss_seidel(m, equations, periods, max_ite)
}
\arguments{
\item{m}{The initialized matrix obtained with \code{.make_matrix()}.}

\item{equations}{Prepared equations with \code{.prep_equations()}.}

\item{periods}{Total number of rows (periods) in the model.}

\item{max_ite}{Maximum number of iterations allowed per block per period.}
}
\description{
Gauss Seidel algorithm
}
\details{
This is the main algorithm of the package. It simulates the model
by recursion with the help of four nested for loops. At each round of
iteration, the values calculated are compared to the previous values. If
the difference is below 1e-4, the round of calculations have converged
and the algorithm jump to the next block of equations.

The algorithm modifies a matrix in place to optimize its performance.
}
\author{
João Macalós
}
