% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_is_hidden_true.R
\name{.sfcr_is_hidden_true}
\alias{.sfcr_is_hidden_true}
\title{Check if hidden equation condition is validated}
\usage{
.sfcr_is_hidden_true(m, hidden)
}
\description{
Check if hidden equation condition is validated
}
