% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_scenario.R
\name{sfcr_scenario}
\alias{sfcr_scenario}
\title{Add scenarios to a \code{sfcr} model.}
\usage{
sfcr_scenario(sfcr_sim, scenario, periods, max_iter = 350)
}
\arguments{
\item{sfcr_sim}{A model generated with the \code{sfcr_sim()} function.}

\item{scenario}{A list containing one or more lists that hold the information
about the scenario(s). Each list must contain three information:
\enumerate{
\item A list describing the variables that are going to be modified and their new
values. This must be written with R equation syntax.
\item An integer defining the period the the scenario starts.
\item An integer defining the period when the scenario ends.
}

See examples for details.}

\item{periods}{The total periods in the model.}

\item{max_iter}{The maximum allowed iteration per period.}
}
\description{
Add scenarios to a \code{sfcr} model.
}
\details{
Add scenario(s) to a model generated with \code{sfcr_sim()} functions.
}
\seealso{
\code{\link{sfcr_sim}}
}
