% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_set.R
\name{sfcr_set}
\alias{sfcr_set}
\title{Define the formulas of the model}
\usage{
sfcr_set(...)
}
\arguments{
\item{...}{The formulas used to define the equations and external
values of the system}
}
\description{
Define the formulas of the model
}
\examples{
sfcr_set(alpha0 ~ 15, alpha1 ~ 0.8, alpha2 ~ 0.15)

}
\author{
João Macalós
}
