% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcr_sim.R
\name{sfcr_sim}
\alias{sfcr_sim}
\title{Simulate a stock-flow consistent model}
\usage{
sfcr_sim(
  equations,
  t = 100,
  exogenous,
  parameters,
  random = NULL,
  initial = NULL,
  hidden = NULL,
  max_iter = 100
)
}
\arguments{
\item{equations}{A list containing all the equations of the model to be simulated.
See details.}

\item{t}{A number specifying the total number of periods of the model to be simulated. It should be at least 2 periods.}

\item{exogenous, parameters, initial}{Named lists with exogenous, parameters, and initial values
of endogenous variables as values. See details.}

\item{random}{Named list with the name of the random component as name, and the desired mean
and standard deviation as values.}

\item{hidden}{Named list that identify the two variables that make the hidden equality
in the SFC model, e.g., \code{list("H_h" = "H_s")}. Defaults to NULL.
If \code{hidden} is supplied, the model will evaluate if the hidden equation is satisfied.
If it is not, it will throw out an error.}

\item{max_iter}{Maximum iterations allowed per period.}
}
\description{
The \code{sfcr_sim} function is used to simulate a SFC model. With adequate number of
periods, the simulated model should converge to a steady state scenario.
}
\details{
The output of  \code{sfcr_sim()} will contain a tibble with the exogenous and endogenous
variables, as well as the parameters, of the simulated model as columns.

The equations at the \code{equations} argument should take the usual R formula syntax.
Also, endogenous and exogenous variables should explicitly indicate if they're
to be considered contemporaneously or with lags, while parameters in should not
be accompanied by period of evaluation indication. For example, a consumption
equation should be written like C_d[t] ~ alpha1 * YD[t] + alpha2 * H_h[t-1].

The named lists should not include time indications. See examples.
}
\examples{
eqs <- list(
  TX_s[t] ~ TX_d[t],
  YD[t] ~ W[t] * N_s[t] - TX_s[t],
  C_d[t] ~ alpha1 * YD[t] + alpha2 * H_h[t-1],
  H_h[t] ~ YD[t] - C_d[t] + H_h[t-1],
  N_s[t] ~ N_d[t],
  N_d[t] ~ Y[t] / W[t],
  C_s[t] ~ C_d[t],
  G_s[t] ~ G_d[t],
  Y[t] ~ C_s[t] + G_s[t],
  TX_d[t] ~ theta * W[t] * N_s[t],
  H_s[t] ~ G_d[t] - TX_d[t] + H_s[t-1]
)

exg <- list("G_d" = 20, "W" = 1)
params <- list("alpha1" = 0.6, "alpha2" = 0.4, "theta" = 0.2)

# t is set to 10 to run faster. A usual model should run at least 50 periods to find a steady state
sfcr_sim(equations = eqs, t = 10, exogenous = exg, parameters = params, initial = NULL)

}
